refactor: Extract the dictionary type to an external file
