+++ scripts/dictionaries.ts
@@ -1,13 +1,9 @@
 import 'server-only'
-import { Locale } from '@/types'
+import type { Locale } from '@/types'
 
-const dictionaries = {
+export const dictionaries = {
   en: () => import('../dictionaries/en.json').then((module) => module.default),
   pl: () => import('../dictionaries/pl.json').then((module) => module.default)
 }
 
 export const getDictionary = async (locale: Locale) => dictionaries[locale]()
-
-export type Dictionary = Awaited<
-  ReturnType<(typeof dictionaries)[keyof typeof dictionaries]>
->
