+++ types/dictionary.tsx
@@ -1,0 +1,5 @@
+import { dictionaries } from '@/scripts'
+
+export type Dictionary = Awaited<
+  ReturnType<(typeof dictionaries)[keyof typeof dictionaries]>
+>
